+++ package.json
@@ -42,6 +42,7 @@
     "babel-minify": "^0.2.0",
     "babel-minify-webpack-plugin": "^0.2.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-minify": "^0.2.0",
     "babel-preset-react": "^6.24.1",
@@ -109,6 +110,7 @@
     "webpack": "^3.5.5",
     "webpack-bundle-analyzer": "^2.9.0",
     "webpack-dev-server": "^2.4.5",
+    "webpack-merge": "^4.1.1",
     "webpack-node-externals": "^1.6.0",
     "webpack-sources": "1.0.1"
   },
